Small fixups
Try to make the code const-correct and make sure functions that shouldn't take parameters take void parameter instead of unspecified number of parameters.
It also bumps the esp-serial-flasher version to solve builds failing for latest versions of ESP-IDF and adds PEDANTIC_FLAGS to the CI to spot such issues in the future.