fix(components/esp-serial-flasher): Fix UART flashing devkits with UART converter with ESP-Prog2
This commit fixes the flashing issue with ESP-Prog2 when using devkits
UART converter. The issue was because pins used for UART were not able
to sink enough current to pull the TX line low. This was fixed in
esp-serial-flasher, so this commit updates the component to use the
latest version.