fix(msc): flash only real payload when UF2_FLAG_MD5_PRESENT flag is set
MD5 data in the UF2 block is not considered, because flashing is done using
16 kB chunks, and it would be necessary to save the MD5 data and then check
the all the MD5 checksums for the chunk. This would be quite time consuming
and would not be very useful, because data arrived through USB have its own
checksum and esp-serial-flasher does MD5 checksum verification over the whole
flash image at the end of the flashing process by default, but can be explicitly
enabled by CONFIG_SERIAL_FLASHER_MD5_ENABLED option.