fix(debug): use correct pins for TDO and TDI
Wrong pins were used when VEND_JTAG_GETTDO was called by JTAG. After this
command, the pin controlled by the target should be read (TDO), the one
controlled by debugger was used instead (TDI). Also when RTS or DTR change
was requested, the TDI was set to log. 0, but the pin is input only. This
is fixed by setting the TDO pin to log. 0 instead.